requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic[email]==2.5.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pyjwt==2.8.0
python-multipart==0.0.6